Add actors folder and DrawObject class
Refactored drawing logic by moving DrawObject and RainingItem classes to a new assets folder for better organization.

Note: Classes from main (Robot, Coin, and Monster) will be moved to a new actors folder in a future commit.
